common/rgw: revet the gmtime() output to exclude the 'T'
This is to revert the changes to gmtime() that were cherry-picked in:
commit 1f9a0e98862112400352a4d31036aec9428e18a8.
The original pupose of the cherry-pick was to resolve rhbz#1958818.

In addition, a new gmtime_iso8601() was added to allow the bucket
notifications timestamp to be ISO 8601 compliant.
This is done, as an alternate resolution to: rhbz#1958818.

(cherry picked from commit 07313a094b455c7834e137d63da57830ef66b672)

Resolves: rhbz#1977271

Signed-off-by: Yuval Lifshitz <ylifshit@redhat.com>

Conflicts:
	src/include/utime.h