rgw: Adding SSE-S3 support in GET and PUT paths (using Vault as KMS)
Added the support to generate KEK based on bucket owner UID in
PutBucketEncryption. This is stored in bucket x-attrs. The KEK-ID is
later used in GET and PUT paths.
In the PUT path, we check if BucketEncryption is enabled for the bucket.
If yes, we detemine if the encryption type is AES256 (i.e., SSE-S3),
then we fetch the KEK-ID from the bucket x-attrs and use it to wrap the
data key. Thereafter, we call generate-data-key. We store the KEK-ID
and the wrapped data-key in the object x-attrs.
In the GET path, we simply pull out the KEK-ID from the object x-attr
and decrypt the object.

Resolves: rhbz#2087986

Signed-off-by: Priya Sehgal <priya.sehgal@flipkart.com>
(cherry picked from commit 7de4479d715ed6b6b16afa0b4c5f316030681593)