rgw: remove unused routine in rgw_admin
Routine "dump_bi_entry" contains logic to decode rgw_bucket_dir_entry
which can contain etags to be fixed.  However, it turns out this
routine is never called, so let's just get rid of it.  (The replacement
code lives in cls land.)

XXX temp proof of concept - should really remove it not #if 0

Fixes: https://tracker.ceph.com/issues/?????

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Resolves: rhbz#1995365