rgw: refactor and revise tag fixes
The first version of this logic attempted to ensure that RGW_ATTR_ETAG
values were always nul terminated strings, which matched the "olden-times"
behavior.  This version attempts to instead make those nuls go away,
which is the new standard.

Also refactored so that the string handling part can be used directly
from more places in radsogw.  Turns out etags can be in rgw_bucket_dir
maps and get passed back as strings from cls code that runs on the osds,
so we need to handle that form also.

Fixes: https://tracker.ceph.com/issues/?????

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Resolves: rhbz#2011456