rgw: "reshard cancel" errors with "invalid argument"
When `radosgw-admin reshard cancel ...` was used on a bucket that is
not currently undergoing resharding, it would error out.

Now it will then try to unschedule any reshards for the bucket in the
so-called "reshard log". It will error out if there is neither an
underway resharding nor a scheduled resharding.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#2027374
(cherry picked from commit f499a98ef1ac1eb7c6e4cf148eae3c41dae23343)
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>