DELETEME: RGWPutBucketInstanceInfoCR crashes invalid layout
We occasionally see layouts of generation 2, with normal index format,
and no logs.

We have been unable to find the source, since it doesn't replicate.

As such we are setting up an assertion to kill the process whenever an
attempt is made to write such a layout, so we can find the source.

Sorry, last attempt topped out at the async processor so we had to
move it higher.

THIS COMMIT ACTIVELY ATTEMPTS TO BREAK CEPH.

IT MUST NOT BE PRESENT IN ANY RELEASED VERSION.

PLEASE REBASE TO REMOVE IT.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>