rgw multisite: resharding scales up shard counts 4x faster
in multisite reshard, we need to keep the old index shards around until
other zones finishing syncing from them. we don't want to allow a
bunch of reshards in a row, because we have to duplicate that many
sets of index objects. so we impose a limit of 4 bilog generations (or 3
reshards), and refuse to reshard again until bilog trimming catches up/
trims the oldest generation

under a sustained write workload, a bucket can fill quickly and need
successive reshards. if we have a limit of 3, we should make them count!
so instead of doubling the shard count at each step, multiply by 8
instead when we're in a multisite configuration

Signed-off-by: Casey Bodley <cbodley@redhat.com>
(cherry picked from commit f3d5a52a81bb0115fe440209b08772a70f1c1b81)

Conflicts:
	src/rgw/rgw_quota.cc
	src/rgw/rgw_quota.h
	src/rgw/rgw_rados.cc
 - Upstream SAL

Resolves: rhbz#2075214
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>