RGW: Add a reader feature
This feature is prompted by a desire to audit contents of
a cluster safely. Currently, such task is done by a privileged
account, which can damage the data by accident. In this
situation, using an account that can only read is an application
of the least privilege.

Note that the reader is a role in RBAC. So, an account can have
both a reader role and a normal user role. The latter allows it to
write the audit report into the cluster, if the operator wants that.

In OpenStack terms, this reader is known as "system reader persona".

Resolves: rhbz#1901857

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
(cherry-picked from commit 991f2431829d21d7a4bffdec90cc3f1afcd4cec5)
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>