rgw: fix http client relevant headers behavior
In a containerized install with keystone; radosgw was caught redhanded
in the act of attempting to validate a uesr supplied token with an empty
admin token.  The problem turned out to be an errant constructor argument
{ fixed in a previous commit } - but the problem was masked by
debug behavior which considered all headers to be relevant causing
the bug to go away.

This fix:
/1/ debug behavior does not alter what are "relevant headers".
/2/ debug behavior produces useful diagnostics for uninteresting headers.
/3/ refactor some of the other logic in receive_headers, which was
    previously recalculating val_loc_s and val_loc_e twice.

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Resolves: rhbz#2140014
(cherry picked from commit a93d0e57290ebf0969a1c55a3074137ac9c67509)

Conflicts:
	src/rgw/rgw_http_client.cc
NB: the "previous" commit mentioned above is in rhceph 5.x; not 6.0