mgr/dashboard: osd form preselect db/wal device filters
If the hostname is selected for the primary devices, then we can
preselect the hostname filter for the db/wal devices because osds will
be deployed only on the hostname of the primary device. If preselected
it'll be clear that only this devices will be used to deploy.

Addition to this, usually ssd devices are used for db/wal devices. So I
am preselecting these too in the filters.

Resolves: rhbz#2101426

Fixes: https://tracker.ceph.com/issues/57118
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit 5f8eff3beff7b8aa0066811dedf0024b76001a6e)
(cherry picked from commit fb897e6ae49105d064bf714182060d6f0424f22c)