rgw/s3: align output format with client request
In general, e.g., in the S3 ListBucket response, the
output Formatter is already of the type requested in the HTTP
Accept header.

Therefore, while there are still many instances where the of
Formatter is assumed to be XML, it appears necessarily safe
and seemingly correct to extend this assumption to end_header(...),
which sends the response content-type.

Fixes: https://tracker.ceph.com/issues/55680
Resolves: rhbz#2028220

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>
(cherry picked from commit 7ad538611094117ef5b5c6d4a71f0414d11c5389)
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>