cls/fifo: Remove all use of 'tags'
A lot of stuff in fifo is overengineered. The 'tags' functionality
requires the client to keep track of a random 'tag' associated with
the head part in order to push.

The problem is that there's nothing this really guards against. Worse,
it turns creating a new part and setting the head (both of which
should be idempotent operations) into operations that /aren't/
idempotent, creating the potential for getting into an
inconsistent (and wedged) state where one wouldn't have existed
otherwise.

Thus, we keep them around so as not to change the serialized format, but
the CLS module no longer test for them and the client no longer tracks
them.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>
(cherry picked from commit f9f0c0990c47d105d976e4de7991e67bd2cf4856)
Resolves: rhbz#2075212
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>