mgr/cephadm: be aware of host's shortname and FQDN
The idea is to gether the shortname and FQDN as part
of gather-facts, and then if we ever try to check if a certain
host is in our internal inventory by hostname, we can check
these other known names. This should avoid issues where
we think a hostname specified by FQDN is not in our
inventory because we know the host by the shortname
or vice versa.

Fixes: https://tracker.ceph.com/issues/58738

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 6443cf15d54a5e50c245dd08c3db005bb8521b6a)

Conflicts:
	src/pybind/mgr/cephadm/tests/test_cephadm.py

Resolves: rhbz#2392390