rgw: add an "object reindex" subcommand to radosgw-admin
This subcommand will take a bucket and an object (or a list of objects
in a file) and add those objects to the bucket's index. It does so by
preparing the bucket index entry update and then allowing the
so-called "dir suggest" mechansim, which is triggered by listing the
bucket, to completing the bucket index entry. This mechanism is
triggered by listing the bucket.  Using this mechanism has the
advantage of doing this lazily that both allows the reindex operation
to run more quickly and distributes the workload over time.

This subcommand can help restore lost bucket index entries.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

(cherry picked from commit be0583c3082b9f924fcdbe4c81d540ad08bcca24)

    Conflicts:
            src/rgw/rgw_rados.cc : reverted zipper datatype updates
            src/rgw/rgw_admin.cc : reverted zipper datatype updates

Resolves: rhbz#2174235