rgw/multisite: handle shard_id correctly in data sync when bucket num_shards is 0
For buckets that have num_shards set to 0, bucket instance
will not have a shard_id delimiter. When this bucket instance is parsed,
we end up assigning a value of -1 to shard_id, which is invalid
in data sync. This change ensures that we represent the shard_id correctly
by giving it a valid number

Resolves rhbz#2188022
Signed-off-by: Shilpa Jagannath <smanjara@redhat.com>
(cherry picked from commit 9aeb1e52285f067c8c20c1846305c38ff234e35b)