rgw: allow object locator to be specified in cls_rgw_bucket_complete_op
When a tag is not specified in bucket operation (i.e., it's not
transactional), the prepare function does not have to be called. But
then there was no way to specify a locator for the object since that
wasn't a parameter in the complete phase. This adds an object locator
parameter to complete.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#2224636