os/bluestore: Introduce reblob_extents
Func reblob_extents is used to modify extents.
It is final step of melding 2 blobs together.
It removes reference to blob that is to be phased out,
and replaces it with reference to the blob that is the sum of them.

Signed-off-by: Adam Kupczyk <akupczyk@ibm.com>
(cherry picked from commit 9d963b8e7c3f8af8b60c031e11325c02b831f445)

Resolves: rhbz#2218013