build: Fix fmt version check
Currently, when attempting to build ceph on a system with fmt
installed, we try to build against it whatever the version. This
constantly breaks people's builds, since newer versions of fmt often
change the API.

This change specifies that versions must be below 10 as well as at or
above 8.1.1, so that on systems with a new format, we fall back to
using the submodule.

It also removes the `Findfmt.cmake` module, as that does not check
the installed version. Instead, we use the cmake config file installed by
the system package of fmt and does support version checking.

Resolves: rhbz#2252256

Signed-off-by: Adam Emerson <aemerson@redhat.com>
(cherry picked from commit 686dd3d838b6de6bf1101f0b61f729426b0bccad)
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>