pybind/mgr: use black & isort on the smb module
Provide tox envs that check or reformat code with black and isort,
currently applied to only the new smb module.
This is similar to what we recently did for enabling tox in the
cephadmlib dir as it only applies to new code. However, other modules
that want to opt-in to automated, python-community-wide typical,
stop-thinking-and-let-tools-do-it approach to code formatting can
be added to the new envs later on.

Resolves: rhbz#2300005
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 3981d3d3841efc0c867c3f6084e4a5522c79e347)