rgw: stand-in fix for complete-multipart return ETag
This version refreshes object attrs after upload completes, which
seems to be the smallest available fix.  Appears to return the correct
value for ETag.

For 7.1, we could replace this fix with a validated backport of (at
least) https://github.com/ceph/ceph/pull/54569.  This is a more intrusive
change and original author should do the backport.

Resolves: rhbz#2266579
Resolves: rhbz#2271117

Signed-off-by: matt benjamin <mbenjamin@redhat.com>